+++ src/api.sc2-directstrike/Controllers/ReplayController.cs
@@ -32,7 +32,7 @@ public class ReplayController : ControllerBase
     }
 
     [HttpGet("{id}")]
-    public async Task<Replay?> GetById(string pkt, int id)
+    public async Task<Replay?> GetById(string pkt, ulong id)
     {
         if (pkt.Length != 24)
         {
@@ -76,7 +76,7 @@ public class ReplayController : ControllerBase
     public static async Task<Replay> GenerateIncrementedReplay(string pkt, PostReplay postReplay, DbContext dbContext)
     {
         Replay replay = postReplay;
-        uint id = await dbContext.WriteToDb(pkt, NAME, replay);
+        ulong id = await dbContext.WriteToDb(pkt, NAME, replay);
 
         return replay with { Id = id };
     }
